dbapi/Cursor: detect and raise an error for unsupported SQL
Cloud Spanner doesn't currently support statements
such as

    ALTER TABLE t ADD CONSTRAINT t_al_m UNIQUE (al, m)

This change adds a pass that if given known unsupported
SQL patterns, will return True or False.

This change will currently raise an Exception but that'll
be dealt with in a follow-up change on whether to log it
to the caller or transform it entirely.

Fixes #37