dbapi,django: generate missing ids + fix SQL alter statements
* Generates the missing id field for columns and generates
corresponding uuid4() for the params/values, and this is an
alright fix because currently the only user of dbapi is
spanner/django, which can generate AutoFields with `primary_key=True`
which assumes that the Database will generate ids automatically.

Therefore, given:

    Columns: ['app', 'name']
    Params:  [('ap', 'n',), ('bp', 'm',)]

It'll get fixed up to

    Columns: ['app', 'name']
    Params:  [
        ('ap', 'n', '8a17c112-fbd1-4c03-90f3-df8212c17f47',),
        ('bp', 'm', '7340b5e1-97f6-44ba-985f-e2d98f32c81e',),
    ]

* Adds the appropriate sql_alter* statements to
spanner/djang/schema.DatabaseWrapper to match Spanner's SQL

* Makes noops for lastrowid, commit, rollback since Spanner
doesn't return the lastrowid and the DBAPI v2 instructs us to
return None if the DB doesn't support lastrowid. commit and rollback
are now just `pass` because the underlying implementation uses
transactions under the hood

Updates #46
Fixes #50