dbapi: use sentinel type TimestampStr type to indicate Spanner.TIMESTAMP
When performing an UPDATE, Cloud Spanner's transaction methods
have two variants:
a) execute_update
b) insert_or_update

where a) supports a full UPDATE with conditions
but b) has only 3 parameters:
* table
* columns
* values

and b) accepts param_types=None for complex types such as
spanner.TIMESTAMP and it invokes the RPC API. As documented
in
    https://cloud.google.com/spanner/docs/data-types#canonical-format_1
the client libraries accept timestamps in the language specific
timestamp format,

WHILE

a) performs and ExecuteSql API request with DML and thus requires
the type to be declared in param_types.

This change depends on its parent PR #81 and will fix the reported
problem, by checking whether our spanner/django/* returned a sentinel
type 'TimestampStr' that inherits from 'str', to indicate that we'll
need to send it as a spanner.TIMESTAMP.

With that change, the repro program:

    cur.execute(
      'UPDATE article SET headline = %s, pub_date = %s '
      'WHERE article.id = %s',
      ('Article #1', Timestamp('2005-08-30T01:01:01.000001Z'), 1954819271609460848))

and since TimestampStr is only used by our internal functions, it is
safe to directly infer spanner.TIMESTAMP from it.

While here, also inferring Spanner.BOOL and Spanner.INT64.

Fixes #82