dbapi: Connection.commit to now batch operations
Knowing the operations permitted by Batch as per

    https://googleapis.dev/python/spanner/latest/batch-usage.html

being:
* insert(table, colums, values)
* update(table, colums, values)

this change makes use of Connection.commit to batch
up such operations.

With this change, running the `basic` Django test suite
time goes from:

    15min 18sec

down to

    6min 16sec

which is a 59.01% reduction in time.

A follow-up PR will further accept batches and transactions
and then continguously group them to make for effective
grouping since some operations can be run on Batch but
not on Transaction.

Updates #142