dbapi: batch continguous CREATEs, except CREATE DATABASE
Batch continguous (in-a-row) CREATE statements
and send them all at once if a non-CREATE like
operation is invoked on when .close() or any iteration
is invoked.

When isolated, this makes for a 95% speed-up, but
when run with Django basic tests, there is no speed-up
noticed, perhaps because after every table is created,
Django invokes the operation to list tables which causes
a flush and basically the older behavior.

Fixes #190